Move `UiLayout` to `re_ui`
