+++ crates/viewer/re_ui/src/lib.rs
@@ -14,6 +14,7 @@ pub mod notifications;
 mod section_collapsing_header;
 pub mod syntax_highlighting;
 mod ui_ext;
+mod ui_layout;
 pub mod zoom_pan_area;
 
 use egui::Color32;
@@ -30,6 +31,7 @@ pub use self::{
     section_collapsing_header::SectionCollapsingHeader,
     syntax_highlighting::SyntaxHighlighting,
     ui_ext::UiExt,
+    ui_layout::UiLayout,
 };
 
 // ---------------------------------------------------------------------------
